test: introduce ChannelSpy helper class for improved test isolation and refactor integration tests to utilize it
