+++ tests/python/unit/test_history_consumer.py
@@ -3,7 +3,8 @@ from types import SimpleNamespace
 
 import pytest
 
-from conftest import ChannelSpy, ConnectionSpy
+from conftest import ConnectionSpy
+from tests.python._helpers import ChannelSpy
 
 
 def test_market_messages_route_to_market_insert(history_consumer_module, monkeypatch):
